requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
drf-writable-nested==0.5.1
html5lib==1.0.1
Pillow==6.1.0
psycopg2==2.8.3
PyPDF2==1.26.0
pytz==2019.2
reportlab==3.5.23
six==1.12.0
sqlparse==0.3.0
webencodings==0.5.1
xhtml2pdf==0.2.3